Refactor: Decouple OCR recognition functionality into separate module
- Created new recognition.ts module with OCR-specific code
- Updated index.ts to import and use the new module
- Modified package.json to correctly reference main entry point
- Re-exported recognition functions for cleaner API access
