+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "ocr",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "ocr": "tsc && node dist/index.js",
