+++ src/index.ts
@@ -1,92 +1,15 @@
-import { Mistral } from '@mistralai/mistralai';
-import fs from 'fs';
 import path from 'path';
+import fs from 'fs';
 import dotenv from 'dotenv';
 import { glob } from 'glob';
+import { 
+  initClient, 
+  processPdfFile, 
+  combineOcrResults, 
+  saveToJson 
+} from './recognition';
+
 dotenv.config();
-
-const apiKey = process.env.MISTRAL_API_KEY;
-const client = new Mistral({apiKey: apiKey});
-
-// Function to save object to JSON file
-function saveToJson(data: any, filename: string, outputDir?: string): void {
-  try {
-    const jsonString = JSON.stringify(data, null, 2);
-    // Use the provided output directory or fall back to current working directory
-    const baseDir = outputDir ? outputDir : process.cwd();
-    const filePath = path.join(baseDir, filename);
-    
-    // Ensure the directory exists
-    const dirname = path.dirname(filePath);
-    if (!fs.existsSync(dirname)) {
-      fs.mkdirSync(dirname, { recursive: true });
-    }
-    
-    fs.writeFileSync(filePath, jsonString);
-    console.log(`Data successfully saved to ${filePath}`);
-  } catch (error) {
-    console.error('Error saving data to JSON file:', error);
-  }
-}
-
-// Process a single PDF file
-async function processPdfFile(filePath: string): Promise<any> {
-  try {
-    const fileContent = fs.readFileSync(filePath);
-    const fileName = path.basename(filePath);
-    
-    const uploadedPdf = await client.files.upload({
-      file: {
-        fileName: fileName,
-        content: fileContent,
-      },
-      purpose: "ocr" as any
-    });
-
-    const signedUrl = await client.files.getSignedUrl({
-      fileId: uploadedPdf.id,
-    });
-
-    const ocrResponse = await client.ocr.process({
-      model: "mistral-ocr-latest",
-      document: {
-        type: "document_url",
-        documentUrl: signedUrl.url,
-      }
-    });
-    
-    return ocrResponse;
-  } catch (error) {
-    console.error(`Error processing file ${filePath}:`, error);
-    throw error;
-  }
-}
-
-// Combine OCR results, adjusting page indices
-function combineOcrResults(results: any[]): any {
-  if (results.length === 0) return { pages: [] };
-  
-  const combined = { ...results[0] }; // Clone the first result
-  let pageOffset = combined.pages.length;
-  
-  // Start from the second result
-  for (let i = 1; i < results.length; i++) {
-    const result = results[i];
-    
-    // Adjust page indices and add to combined result
-    const adjustedPages = result.pages.map((page: any) => {
-      return {
-        ...page,
-        index: page.index + pageOffset
-      };
-    });
-    
-    combined.pages = combined.pages.concat(adjustedPages);
-    pageOffset += result.pages.length;
-  }
-  
-  return combined;
-}
 
 /**
  * Process PDF files and extract text using OCR
@@ -107,6 +30,9 @@ async function processOcr(
       return;
     }
     
+    // Initialize client once to be reused
+    const client = initClient();
+    
     // If in single file mode, only process the first file
     const filesToProcess = singleFileMode ? [pdfFiles[0]] : pdfFiles;
     const ocrResults = [];
@@ -114,7 +40,7 @@ async function processOcr(
     // Process each PDF file
     for (const file of filesToProcess) {
       console.log(`Processing ${file}...`);
-      const result = await processPdfFile(file);
+      const result = await processPdfFile(file, client);
       ocrResults.push(result);
       
       // Save individual result with custom output directory
@@ -263,6 +189,13 @@ async function main() {
 if (require.main === module) {
   main();
 } else {
-  // Export the function for use in other modules
-  module.exports = { processOcr };
+  // Export the functions for use in other modules
+  module.exports = { 
+    processOcr,
+    // Re-export recognition functions to provide a clean API
+    processPdfFile,
+    combineOcrResults,
+    saveToJson,
+    initClient
+  };
 }
